requirements.txt
sentence-transformers
transformers
torch
scikit-learn==1.3.2
numpy
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
